requirements.txt
Django==3.2.2
django-versatileimagefield==2.0
xhtml2pdf==0.2.5
